requirements.txt
flask
pandas
scikit-surprise
werkzeug
numpy<2.0
